🔨 categorical data should hide that it is dict/ordinal encode
Note that this will only be abled after
df = df._auto_encode()

which should be the default in version 5

+++ tests/agg_test.py
@@ -472,7 +472,7 @@ def test_format_xarray_and_list(df_local):
     assert isinstance(count, list)
 
     df = df[:3]
-    df['g'] = df.x
+    df['g'] = df.x.astype('int32')
     df.categorize(df.g, labels=['aap', 'noot', 'mies'], inplace=True)
     count = df.count(binby='g', array_type='xarray')
     assert count.coords['g'].data.tolist() == ['aap', 'noot', 'mies']
